KITI-2980
Enhancement to prevent a vendor that has a B2B contract from having no default commodity code defined.  If a requisition is created from an E-Shop order with invalid commodity codes, the default commodity code for that vendor will be assigned to any problematic line item.  Also prevents addition of duplicate commodity codes and ensures that any commodity code that is added to a vendor through hand keying the code into the field will be validated against the commodity code table.