KFSPTS-764: The current implementation deals with both foreign and domestic wire charges. These wire charge amounts are maintained via the Disbursement Voucher Wire Charge maintenance table. Either one OR both of these values (domestic vs foreign) could be zero. The wire charge value that is used (domestic vs foreign) is based upon the disbursement voucher bank country code. Domestic is used when that country code is US; otherwise, foreign is used.
In order to prevent the zero dollar general ledger pending entries from being created, a verification check needs to be added that looks at the disbursement voucher bank country code as well as the wire charge type (domestic vs foreign) associated with the individual DV being processed.

When the wire charge amount is found to be zero dollars based on the bank country code, do NOT create any general ledger pending entries for the wire charge; otherwise create the general ledger pending entries for the wire charge. 

Note: The General Ledger Pending Entries associated with the DV and the bank account will continue to be generated.