KFSPTS-1710: In Inquiry Section Definition "Asset Detail Information", changed attribute name from inventoryStatus.inventoryStatusCode to inventoryStatusCode so that lookup table would automatically present the descripiton next to the code.
Version 6413 of this file had JIRA queue typed incorrectly so that change was not being tracked on this JIRA number.  The commit comment is listed here again for documentation purposes:
Changed attribute definitions for Asset Status Code, Asset Type Code, Building Code, and Building Room Number from TextControlDefinition to SelectControlDefinition so that the descriptions would appear with the codes on the eDocs for the end users.