KFSUPGRADE-240: Getting Access Security to work
First I removed all the com/rsmart DD refs as we are using the foundation delivered ones now.
Then had to add a bean securityBalanceLookupableHelperService as it is only included in the GL
bean overrides which we dont use.  Then added a hack to keep Access Security from spewing garbage
to the log file.